Add Pillager Patrol Spawn Changes
Pillager patrols now follow the config values we created earlier. Mod compatibility for the spawning chance is interesting. Currently, If we do not get the expected vanilla spawn rate we will let the other mods control it. In future, we should map the modified value to our range then modify it accordingly.
Also added some debug messages so testing can be done on a production server. The code for vanilla pillager patrol spawning is not the easiest to mixin to say the least.