Storage UUIDs, not players
Fixed a bug where if a user relogs, the lastPlayer seems to be invalidated and not used properly when checking for subsequent spawns. We now store the UUID of the last targeted player to get around this.
Also added a debug statement stating the other targets for a patrol/trader if a player is exempt from one.